refactor common.js as commands
